+++ fixtures/coverall/tests/bindings/test_coverall.py
@@ -75,7 +75,7 @@ class TestCoverall(unittest.TestCase):
         with self.assertRaises(CoverallError.TooManyHoles):
             Coveralls.fallible_new("", True)
 
-        with self.assertRaisesRegex(InternalError, "expected panic: woe is me"):
+        with self.assertRaisesRegex(InternalError, "Rust panic"):
             Coveralls.panicing_new("expected panic: woe is me")
 
         # in the absence of cycles Python is deterministic in killing refs
@@ -85,15 +85,35 @@ class TestCoverall(unittest.TestCase):
         self.assertEqual(get_num_alive(), 0)
 
 
-    def test_errors(self):
+    def test_simple_errors(self):
         coveralls = Coveralls("test_errors")
         self.assertEqual(coveralls.get_name(), "test_errors")
 
         with self.assertRaises(CoverallError.TooManyHoles):
             coveralls.maybe_throw(True)
 
-        with self.assertRaisesRegex(InternalError, "expected panic: oh no"):
+        with self.assertRaisesRegex(InternalError, "Rust panic"):
             coveralls.panic("expected panic: oh no")
+
+    def test_complex_errors(self):
+        coveralls = Coveralls("test_complex_errors")
+
+        # Test success
+        self.assertEqual(True, coveralls.maybe_throw_complex(0))
+
+        # Test errors
+        with self.assertRaises(ComplexError.OsError) as cm:
+            coveralls.maybe_throw_complex(1)
+        self.assertEqual(cm.exception.code, 10)
+        self.assertEqual(cm.exception.extended_code, 20)
+
+        with self.assertRaises(ComplexError.PermissionDenied) as cm:
+            coveralls.maybe_throw_complex(2)
+        self.assertEqual(cm.exception.reason, "Forbidden")
+
+        # Test panics, which should cause InternalError to be raised
+        with self.assertRaises(InternalError) as cm:
+            coveralls.maybe_throw_complex(3)
 
     def test_self_by_arc(self):
         coveralls = Coveralls("test_self_by_arc")
@@ -116,7 +136,7 @@ class TestCoverall(unittest.TestCase):
         with self.assertRaises(CoverallError.TooManyHoles):
             coveralls.take_other_fallible()
 
-        with self.assertRaisesRegex(InternalError, "expected panic: with an arc!"):
+        with self.assertRaisesRegex(InternalError, "Rust panic"):
             coveralls.take_other_panic("expected panic: with an arc!")
 
         coveralls.take_other(None)